feat: request に含まれる email で照合して、OKなら削除を許可する機能を追加

+++ app/api/item/delete/[id]/route.js
@@ -1,20 +1,26 @@
 import { NextResponse } from "next/server";
-
 import connectDB from "../../../../utils/database";
-
 import {ItemModel} from "../../../../utils/schemaModels";
 
 export async function DELETE(request,context) {
 
     // console.log(context.params.id);
 
+    const reqBody = await request.json();
+
     try{
         await connectDB();
 
-        await ItemModel.deleteOne({_id: context.params.id});
+        const singleItem =  await ItemModel.findById(context.params.id);
 
+        if(singleItem.email === reqBody.email){
+            await ItemModel.deleteOne({_id: context.params.id});
 
-        return NextResponse.json({message: "アイテム削除成功"});
+            return NextResponse.json({message: "アイテム削除成功"});
+        }else{
+            return NextResponse.json({message: "他の人が作成したアイテムです"});
+        }
+
     }catch{
         return NextResponse.json({message: "アイテム削除失敗"});
     }
